vikasana-api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
asyncpg==0.29.0
alembic==1.13.3
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
email-validator==2.2.0
python-multipart==0.0.9
httpx==0.27.0
minio==7.2.5
itsdangerous==2.1.2
qrcode[pil]==7.4.2
reportlab==4.1.0
pypdf==4.0.1
opencv-contrib-python-headless==4.9.0.80
numpy>=1.26.0
Pillow>=10.0.0
anyio>=4.0.0
redis>=5.0.0
celery[redis]>=5.3.0
